refactor(interactivity): move BehaviorManager into src/interactivity
Pure relocation + barrel. No behaviour change.
