+++ src/interactivity/index.ts
@@ -1,0 +1,3 @@
+// Public interactivity surface. Consumers import from './interactivity'
+// rather than reaching into individual modules.
+export { BehaviorManager, IHtmlBehaviorOptions } from './behavior';
